fix: 🐛 优惠券根据所选服务查询2

+++ carservice-wx-api/src/main/java/org/click/carservice/wx/model/cart/result/CartCheckoutResult.java
@@ -49,7 +49,7 @@ public class CartCheckoutResult implements Serializable {
     /**
      * 优惠券可用数量
      */
-    private Integer availableCouponLength;
+    private long availableCouponLength;
     /**
      * 选中商品金额
      */
